refactor(web): index.ts 추 가 및 import 경로 정리
