+++ apps/web/src/features/space-create/index.ts
@@ -1,1 +1,2 @@
 export { SpaceCreateForm } from "./ui/space-create-form";
+export { ThemeColorSelect } from "./ui/theme-color-select";
